requirements.txt
Pillow~=11.0.0
geopy~=2.4.1
piexif~=1.1.3
Flask~=3.1.0
Flask-SocketIO~=5.5.1
python-dotenv~=1.0.0
requests~=2.31.0